app for oidc

+++ oscm-ui-tests/src/test/java/org/oscm/app/AppAzureControllerWT.java
@@ -47,47 +47,69 @@ public class AppAzureControllerWT {
 
   @Test
   public void test01setSettingsIntoController() throws Exception {
-    controllerTester.changeValueInputInSpecificField("49:1", 56, changedUserID);
-    controllerTester.changeValueInputInSpecificField("49:2", 56, userKey);
-    controllerTester.changeValueInputInSpecificField("49:3", 55, changedPassword);
+    if (controllerTester.getAuthenticationMode().equals("OIDC")) {
+      controllerTester.log("OIDC MODE SKIPPING TEST");
+    } else {
 
-    controllerTester.buttonClickEvent(75);
+      controllerTester.changeValueInputInSpecificField("49:1", 56, changedUserID);
+      controllerTester.changeValueInputInSpecificField("49:2", 56, userKey);
+      controllerTester.changeValueInputInSpecificField("49:3", 55, changedPassword);
 
-    assertEquals(changedUserID, controllerTester.readValue("49:1", 56));
-    assertEquals(userKey, controllerTester.readValue("49:2", 56));
-    assertEquals(changedPassword, controllerTester.readValue("49:3", 55));
+      controllerTester.buttonClickEvent(75);
+
+      assertEquals(changedUserID, controllerTester.readValue("49:1", 56));
+      assertEquals(userKey, controllerTester.readValue("49:2", 56));
+      assertEquals(changedPassword, controllerTester.readValue("49:3", 55));
+    }
   }
 
   @Test
   public void test02changeOrganizationIdInAppConfigurator() throws Exception {
-    tester = new AppTester();
-    tester.loginAppConfig(userID, userPassword);
+    if (controllerTester.getAuthenticationMode().equals("OIDC")) {
+      controllerTester.log("OIDC MODE SKIPPING TEST");
+    } else {
+      tester = new AppTester();
+      tester.loginAppConfig(userID, userPassword);
 
-    tester.changeOrganizationID("//input[@id='configurationSettings:j_idt52:1:configurationValue']", PlaygroundSuiteTest.supplierOrgId);
+      tester.changeOrganizationID(
+          "//input[@id='configurationSettings:j_idt52:1:configurationValue']",
+          PlaygroundSuiteTest.supplierOrgId);
 
-    assertEquals(PlaygroundSuiteTest.supplierOrgId, tester.readDefaultValue("//input[@id='configurationSettings:j_idt52:1:configurationValue']"));
+      assertEquals(
+          PlaygroundSuiteTest.supplierOrgId,
+          tester.readDefaultValue(
+              "//input[@id='configurationSettings:j_idt52:1:configurationValue']"));
+    }
   }
 
   @Test
   public void test03checkThatOrganizationIdIsChanged() throws LoginException, InterruptedException {
-    controllerTester.getDriver().manage().deleteAllCookies();
+    if (controllerTester.getAuthenticationMode().equals("OIDC")) {
+      controllerTester.log("OIDC MODE SKIPPING TEST");
+    } else {
+      controllerTester.getDriver().manage().deleteAllCookies();
 
-    controllerTester.loginAppController(
-        changedUserID, changedPassword, AppPathSegments.APP_PATH_CONTROLLER_AZURE);
+      controllerTester.loginAppController(
+          changedUserID, changedPassword, AppPathSegments.APP_PATH_CONTROLLER_AZURE);
 
-    assertEquals(PlaygroundSuiteTest.supplierOrgId, controllerTester.readValue("49:0", 56));
+      assertEquals(PlaygroundSuiteTest.supplierOrgId, controllerTester.readValue("49:0", 56));
+    }
   }
 
   @Test
   public void test04undoSettingsIntoController() throws Exception {
-    controllerTester.changeValueInputInSpecificField("49:1", 56, userID);
-    controllerTester.changeValueInputInSpecificField("49:2", 56, "1000");
-    controllerTester.changeValueInputInSpecificField("49:3", 55, userPassword);
+    if (controllerTester.getAuthenticationMode().equals("OIDC")) {
+      controllerTester.log("OIDC MODE SKIPPING TEST");
+    } else {
+      controllerTester.changeValueInputInSpecificField("49:1", 56, userID);
+      controllerTester.changeValueInputInSpecificField("49:2", 56, "1000");
+      controllerTester.changeValueInputInSpecificField("49:3", 55, userPassword);
 
-    controllerTester.buttonClickEvent(76);
+      controllerTester.buttonClickEvent(76);
 
-    assertEquals(changedUserID, controllerTester.readValue("49:1", 56));
-    assertEquals(userKey, controllerTester.readValue("49:2", 56));
-    assertEquals(changedPassword, controllerTester.readValue("49:3", 55));
+      assertEquals(changedUserID, controllerTester.readValue("49:1", 56));
+      assertEquals(userKey, controllerTester.readValue("49:2", 56));
+      assertEquals(changedPassword, controllerTester.readValue("49:3", 55));
+    }
   }
 }
